requirements.txt
Flask==2.2.2
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-aiplatform==1.60.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.4
google-cloud-storage==2.18.0
google-cloud-texttospeech==2.17.2
google-generativeai==0.8.3
werkzeug==2.2.2
gunicorn